+++ mitra_models/src/filter_rules/queries.rs
@@ -45,6 +45,7 @@ pub async fn remove_filter_rule(
     Ok(())
 }
 
+/// Returns filter rules in the order of precedence
 pub async fn get_filter_rules(
     db_client: &impl DatabaseClient,
 ) -> Result<Vec<FilterRule>, DatabaseError> {