+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Don't pollute PostgreSQL log with `activitypub_object_pkey` constraint violation errors.
 - Improved error reporting during actor document parsing.
 - Optimized post queries.
+- Improved documentation of `list-filter-rules` command.
 
 ## [3.7.1] - 2024-10-17